pyproject.toml
[project]
name = "medsinyal"
version = "0.1.0"
description = "Pengolahan Sinyal Medis — Biomedical Signal Processing course materials (Universitas Indonesia)"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [
    { name = "Pengolahan Sinyal Medis Team" },
]

dependencies = [
    "numpy>=1.24",
    "scipy>=1.11",
    "matplotlib>=3.7",
    "wfdb>=4.1",
    "mne>=1.5",
    "scikit-learn>=1.3",
    "scikit-image>=0.21",
    "pydicom>=2.4",
    "Pillow>=10.0",
    "pandas>=2.0",
    "jupyter>=1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "nbconvert>=7.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/medsinyal"]

[tool.pytest.ini_options]
testpaths = ["tests"]
